pytest_micro/requirements.txt
flask
pytest
mysql-connector-python
requests